set the ids in pasted features
set the ids in the pasted features instead of in the copied.
with the signal committedFeatureAdded the added_geom receives the copied features instead of the inserted one.
thats why we get the inserted features with the fid of the copied ones.

this hasn't been a problem with copied features with the same id-fields like the pasted ones. but if there is no valid id-field on the source-layer, we cannot set the id in the copied features-